+++ Tests/GRDBTests/RecordPrimaryKeySingleTests.swift
@@ -277,21 +277,19 @@ class RecordPrimaryKeySingleTests: GRDBTestCase {
             try record2.insert(db)
             
             do {
-                // This method used to return nil, but does no longer.
-                // In GRDB 2.0, it won't return an optional.
                 let cursor = try Pet.fetchCursor(db, keys: [])
-                try XCTAssertNil(cursor!.next())
+                try XCTAssertNil(cursor.next())
             }
             
             do {
-                let cursor = try Pet.fetchCursor(db, keys: [["UUID": record1.UUID], ["UUID": record2.UUID]])!
+                let cursor = try Pet.fetchCursor(db, keys: [["UUID": record1.UUID], ["UUID": record2.UUID]])
                 let fetchedRecords = try [cursor.next()!, cursor.next()!]
                 XCTAssertEqual(Set(fetchedRecords.map { $0.UUID! }), Set([record1.UUID!, record2.UUID!]))
                 XCTAssertTrue(try cursor.next() == nil) // end
             }
             
             do {
-                let cursor = try Pet.fetchCursor(db, keys: [["UUID": record1.UUID], ["UUID": nil]])!
+                let cursor = try Pet.fetchCursor(db, keys: [["UUID": record1.UUID], ["UUID": nil]])
                 let fetchedRecord = try cursor.next()!
                 XCTAssertEqual(fetchedRecord.UUID!, record1.UUID!)
                 XCTAssertTrue(try cursor.next() == nil) // end
@@ -351,15 +349,13 @@ class RecordPrimaryKeySingleTests: GRDBTestCase {
             
             do {
                 let UUIDs: [String] = []
-                // This method used to return nil, but does no longer.
-                // In GRDB 2.0, it won't return an optional.
                 let cursor = try Pet.fetchCursor(db, keys: UUIDs)
-                try XCTAssertNil(cursor!.next())
+                try XCTAssertNil(cursor.next())
             }
             
             do {
                 let UUIDs = [record1.UUID!, record2.UUID!]
-                let cursor = try Pet.fetchCursor(db, keys: UUIDs)!
+                let cursor = try Pet.fetchCursor(db, keys: UUIDs)
                 let fetchedRecords = try [cursor.next()!, cursor.next()!]
                 XCTAssertEqual(Set(fetchedRecords.map { $0.UUID! }), Set(UUIDs))
                 XCTAssertTrue(try cursor.next() == nil) // end